+++ .idea/misc.xml
@@ -3,7 +3,7 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Users/zapat/OneDrive/Escritorio/SEMESTRE 6/construccion_proyecto/ProyectoAndroid/app/src/main/res/layout/activity_after_login.xml" value="0.25" />
+        <entry key="..\:/Users/zapat/OneDrive/Escritorio/SEMESTRE 6/construccion_proyecto/ProyectoAndroid/app/src/main/res/layout/activity_after_login.xml" value="0.5" />
         <entry key="..\:/Users/zapat/OneDrive/Escritorio/SEMESTRE 6/construccion_proyecto/ProyectoAndroid/app/src/main/res/layout/activity_main.xml" value="0.25" />
         <entry key="..\:/Users/zapat/OneDrive/Escritorio/SEMESTRE 6/construccion_proyecto/ProyectoAndroid/app/src/main/res/layout/fragment_button_four.xml" value="0.13541666666666666" />
         <entry key="..\:/Users/zapat/OneDrive/Escritorio/SEMESTRE 6/construccion_proyecto/ProyectoAndroid/app/src/main/res/layout/fragment_button_one.xml" value="0.20416666666666666" />
