+++ app/src/main/java/com/example/myapplication/MainActivity.kt
@@ -5,43 +5,24 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.provider.AlarmClock.EXTRA_MESSAGE
 import android.view.View
+import android.widget.Button
 import android.widget.EditText
+import android.widget.TextView
 import androidx.fragment.app.Fragment
 import com.example.myapplication.databinding.ActivityAfterLoginBinding
 import com.example.myapplication.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
 
-    lateinit var binding: ActivityAfterLoginBinding
+
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityAfterLoginBinding.inflate(layoutInflater)
-        setContentView(binding.root)
+        setContentView(R.layout.activity_main)
 
-        binding.button2.setOnClickListener{
-            replaceFragment(FragmentButtonOne())
-        }
-
-        binding.button3.setOnClickListener {
-            replaceFragment(FragmentButtonTwo())
-        }
-
-        binding.button4.setOnClickListener {
-            replaceFragment(FragmentButtonThree())
-        }
-
-        binding.button5.setOnClickListener {
-            replaceFragment(FragmentButtonFour())
-        }
     }
 
-    private fun replaceFragment(fragment : Fragment){
-        val fragmentManager = supportFragmentManager
-        val fragmentTransaction = fragmentManager.beginTransaction()
-        fragmentTransaction.replace(R.id.fragmentContainer, fragment)
-        fragmentTransaction.commit()
-    }
+
 
     fun submitButton(view : View){
         val editText = findViewById<EditText>(R.id.editTextTextName)
@@ -51,4 +32,6 @@ class MainActivity : AppCompatActivity() {
         }
         startActivity(intent)
     }
+
+
 }
